+++ src/patch-revert.sh
@@ -13,9 +13,17 @@ OUTDIR=/mnt/out/android-$ARG_ANDROID_VERSION-$ARCH
 mkdir -p $OUTDIR
 
 cd /mnt/work
-mkdir -p /mnt/gen/out.$ARCH
-rm out ||:
-ln -s /mnt/gen/out.$ARCH/ out
+if [[ "$ARG_NO_GEN" == "false" ]]; then
+  if [[ -e out ]] && [[ ! -L out ]]; then
+    echo "The 'out' directory already exists and is not a symlink."
+    exit 1
+  fi
+  mkdir -p /mnt/gen/out.$ARCH
+  if [[ -e out ]]; then
+    rm out
+  fi
+  ln -s /mnt/gen/out.$ARCH/ out
+fi
 
 # ==============================================================================
 # Apply patches
